修复日志重复写的bug
